net: reorganize crate in anticipation of #1264
Space is made to add `tcp`, `udp`, `uds`, ... modules
