+++ tokio-net/src/util/mod.rs
@@ -1,0 +1,5 @@
+//! Utilities for implementing networking types.
+
+mod poll_evented;
+
+pub use self::poll_evented::PollEvented;
