+++ tokio-tcp/src/stream.rs
@@ -4,7 +4,8 @@ use crate::split::{
 };
 
 use tokio_io::{AsyncRead, AsyncWrite};
-use tokio_net::{Handle, PollEvented};
+use tokio_net::driver::Handle;
+use tokio_net::util::PollEvented;
 
 use bytes::{Buf, BufMut};
 use futures_core::ready;
@@ -125,7 +126,7 @@ impl TcpStream {
     ///
     /// ```no_run
     /// use tokio::net::TcpStream;
-    /// use tokio_net::Handle;
+    /// use tokio_net::driver::Handle;
     ///
     /// # fn dox() -> std::io::Result<()> {
     /// let std_stream = std::net::TcpStream::connect("127.0.0.1:34254")?;
@@ -801,7 +802,7 @@ impl TryFrom<TcpStream> for mio::net::TcpStream {
 
     /// Consumes value, returning the mio I/O object.
     ///
-    /// See [`tokio_net::PollEvented::into_inner`] for more details about
+    /// See [`tokio_net::util::PollEvented::into_inner`] for more details about
     /// resource deregistration that happens during the call.
     fn try_from(value: TcpStream) -> Result<Self, Self::Error> {
         value.io.into_inner()
