+++ tokio-uds/src/datagram.rs
@@ -1,4 +1,5 @@
-use tokio_net::{Handle, PollEvented};
+use tokio_net::driver::Handle;
+use tokio_net::util::PollEvented;
 
 use futures_core::ready;
 use futures_util::future::poll_fn;
@@ -200,7 +201,7 @@ impl TryFrom<UnixDatagram> for mio_uds::UnixDatagram {
 
     /// Consumes value, returning the mio I/O object.
     ///
-    /// See [`tokio_net::PollEvented::into_inner`] for more details about
+    /// See [`tokio_net::util::PollEvented::into_inner`] for more details about
     /// resource deregistration that happens during the call.
     fn try_from(value: UnixDatagram) -> Result<Self, Self::Error> {
         value.io.into_inner()
