+++ tokio-uds/src/stream.rs
@@ -5,7 +5,8 @@ use crate::split::{
 use crate::ucred::{self, UCred};
 
 use tokio_io::{AsyncRead, AsyncWrite};
-use tokio_net::{Handle, PollEvented};
+use tokio_net::driver::Handle;
+use tokio_net::util::PollEvented;
 
 use bytes::{Buf, BufMut};
 use futures_core::ready;
@@ -131,7 +132,7 @@ impl TryFrom<UnixStream> for mio_uds::UnixStream {
 
     /// Consumes value, returning the mio I/O object.
     ///
-    /// See [`tokio_net::PollEvented::into_inner`] for more details about
+    /// See [`tokio_net::util::PollEvented::into_inner`] for more details about
     /// resource deregistration that happens during the call.
     fn try_from(value: UnixStream) -> Result<Self, Self::Error> {
         value.io.into_inner()
