+++ tokio/src/runtime/threadpool/builder.rs
@@ -1,7 +1,7 @@
 use super::{background, Inner, Runtime};
-use crate::reactor::Reactor;
 
 use tokio_executor::threadpool;
+use tokio_net::driver::{self, Reactor};
 use tokio_timer::clock::{self, Clock};
 use tokio_timer::timer::{self, Timer};
 
@@ -343,7 +343,7 @@ impl Builder {
             .around_worker(move |w| {
                 let index = w.id().to_usize();
 
-                let _reactor = tokio_net::set_default(&reactor_handles[index]);
+                let _reactor = driver::set_default(&reactor_handles[index]);
                 clock::with_default(&clock, || {
                     let _timer = timer::set_default(&timer_handles[index]);
                     trace::dispatcher::with_default(&dispatch, || {
